separate quant block and layer + edit the name of classes
